adding logs and status to execution && app verification

+++ EZSave/Models/StatusModel.cs
@@ -6,10 +6,10 @@
         public string SourceFilePath { get; set; } = "";
         public string TargetFilePath { get; set; } = "";
         public string State { get; set; } = "";
-        public int TotalFilesSize { get; set; }
+        public long TotalFilesSize { get; set; }
         public int TotalFilesToCopy { get; set; }
         public int Progression { get; set; }
         public int FilesLeftToCopy { get; set; }
-        public int FilesSizeLeftToCopy { get; set; }
+        public long FilesSizeLeftToCopy { get; set; }
     }
 }
